add jonah's comments
