+++ src/parthenon/package.hpp
@@ -30,6 +30,7 @@
 #include <pack/make_pack_descriptor.hpp>
 #include <pack/make_swarm_pack_descriptor.hpp>
 #include <pack/pack_descriptor.hpp>
+#include <pack/scratch_variables.hpp>
 #include <pack/sparse_pack.hpp>
 #include <pack/swarm_pack.hpp>
 #include <parameter_input.hpp>
